service tests now running; separeted index from service;
