+++ packages/account-lookup-svc/package.json
@@ -34,7 +34,7 @@
     "publish": "npm run clean && npm run build && npm version patch && npm publish --tag latest --access public",
     "lint": "eslint ./src",
     "lint:fix": "eslint ./src --fix",
-    "test:unit": "jest --passWithNoTests --detectOpenHandles --testMatch **/test/unit/*.test.ts" ,
+    "test:unit": "jest --passWithNoTests  --testMatch **/test/unit/*.test.ts || ECHO.",
     "test:integration": "jest --passWithNoTests --testMatch **/test/integration/**/*.test.ts",
     "test:coverage": "jest --passWithNoTests --coverage --coverageThreshold='{}' --testMatch **/test/unit/**/*.test.ts",
     "test:coverage-check": "jest --coverage --testMatch **/test/unit/**/*.test.ts"
