+++ packages/account-lookup-svc/test/unit/mocks/memory_participant_service.ts
@@ -50,10 +50,10 @@ import {ILogger} from "@mojaloop/logging-bc-public-types-lib";
 	) {
 		this.logger = logger;
 	}
-	 getParticipantsInfo(fspIds: string[]): Promise<IParticipant[]> {
+	 async getParticipantsInfo(fspIds: string[]): Promise<IParticipant[]> {
 		 throw new Error("Method not implemented.");
 	 }
-     getParticipantInfo(fspId: string): Promise<IParticipant | null> {
+     async getParticipantInfo(fspId: string): Promise<IParticipant | null> {
          throw new Error("Method not implemented.");
      }
